Splitting up battlehill.cpp
