+++ src/led.hpp
@@ -1,0 +1,17 @@
+/*
+ * led.hpp
+ *
+ *  Created on: Dec 30, 2015
+ *      Author: Joshua Southerland
+ */
+
+#ifndef SRC_LED_HPP_
+#define SRC_LED_HPP_
+
+
+void config_led();
+
+// TODO: move/remove
+void blink_led();
+
+#endif /* SRC_LED_HPP_ */
